wip: show project configuration action
